Development

+++ src/App.Services.ProductAPI/Controllers/v1/ProductAPIController.cs
@@ -1,36 +1,36 @@
 
 using System.Net;
+using Microsoft.AspNetCore.Mvc;
 using App.Services.ProductAPI.Data;
+using Microsoft.EntityFrameworkCore;
 using App.Services.ProductAPI.Models;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 namespace App.Services.ProductAPI.Controllers.v1
 {
-    //[Route("api/v{major:apiVersion}.{minor:apiVersion}.{patch:int}/products")]
-    [Route("api/v{version:apiVersion}/product")]
-    [ApiController]
     [ApiVersion("1.0")]
+    [ApiController]
+    [Route("api/v{version:apiVersion}/product")]
+    //[Route("api/v{major:apiVersion}.{minor:apiVersion}.{patch:int}/products")]
     public class ProductAPIController : Controller
     {
-        private readonly ILogger<ProductAPIController> _logger;
         private Response _response;
         private readonly ApplicationDbContext _dbContext;
+        // private readonly ILogger<ProductAPIController> _logger;
 
         public ProductAPIController(
-                                    ApplicationDbContext dbContext,
-                                    ILogger<ProductAPIController> logger
+                                    ApplicationDbContext dbContext
+                                    // ILogger<ProductAPIController> logger
                                     )
         {
-            _response = new Response();
+            //   _logger = logger;
             _dbContext = dbContext;
-            _logger = logger;
+            _response = new Response();
         }
 
         [HttpGet]
         public async Task<ActionResult<Response>> Get()
         {
-            _logger.LogInformation("orther service call");
+            // _logger.LogInformation("orther service call");
             try
             {
                 IEnumerable<Product> products = await _dbContext.Products
@@ -39,16 +39,14 @@ namespace App.Services.ProductAPI.Controllers.v1
                                                             .Include(p => p.Color)
                                                             .Include(p => p.Brand)
                                                             .ToListAsync();
-                _response.StatusCode = HttpStatusCode.OK;
                 _response.Result = products;
-                _response.IsSuccess = true;
             }
             catch (Exception ex)
             {
                 _response.IsSuccess = false;
                 _response.Message = ex.Message;
                 _response.StatusCode = HttpStatusCode.NotFound;
-                _logger.LogError(ex.Message);
+                //  _logger.LogError(ex.Message);
             }
             return _response;
         }
@@ -58,15 +56,24 @@ namespace App.Services.ProductAPI.Controllers.v1
         {
             try
             {
-                var product = await _dbContext.Products
-                                                    .Include(p => p.Category)
-                                                    .Include(p => p.Size)
-                                                    .Include(p => p.Color)
-                                                    .Include(p => p.Brand)
-                                                    .FirstOrDefaultAsync(p => p.Id == id);
-                _response.StatusCode = HttpStatusCode.OK;
-                _response.Result = product;
-                _response.IsSuccess = true;
+                var product = _dbContext.Products.FirstOrDefault(p => p.Id == id);
+                if (product == null)
+                {
+                    _response.StatusCode = HttpStatusCode.NotFound;
+                    _response.IsSuccess = false;
+                }
+                else
+                {
+                    product = await _dbContext.Products
+                                .Where(p => p.Id == id)
+                                .Include(p => p.Category)
+                                .Include(p => p.Size)
+                                .Include(p => p.Color)
+                                .Include(p => p.Brand)
+                                .FirstOrDefaultAsync();
+
+                    _response.Result = product;
+                }
             }
             catch (Exception ex)
             {
@@ -125,10 +132,7 @@ namespace App.Services.ProductAPI.Controllers.v1
                     totalPages = totalPages,
                     currentPage = currentPage
                 };
-
-                _response.StatusCode = HttpStatusCode.OK;
                 _response.Result = pagination;
-                _response.IsSuccess = true;
             }
             catch (Exception ex)
             {
@@ -139,65 +143,72 @@ namespace App.Services.ProductAPI.Controllers.v1
             return _response;
         }
 
-        [HttpPost("Create")]
         // [Authorize(Roles = "ADMIN")]
-
+        [HttpPost("Create")]
         public async Task<ActionResult<Response>> Create(Product product)
         {
             try
             {
-                // // Xử lý thêm sản phẩm, cho phép các trường Size, Color, Brand và Category nullable
-                // if (ModelState.IsValid)
-                // {
-                //     product.Id = Guid.NewGuid().ToString();
-
-                //     // Nếu người dùng không chọn, các thuộc tính này sẽ là null
-                //     product.Size = product.Size?.Id != null ? context.Sizes.Find(product.Size.Id) : null;
-                //     product.Color = product.Color?.Id != null ? context.Colors.Find(product.Color.Id) : null;
-                //     product.Brand = product.Brand?.Id != null ? context.Brands.Find(product.Brand.Id) : null;
-                //     product.Category = product.Category?.Id != null ? context.Categories.Find(product.Category.Id) : null;
-
-                //     context.Products.Add(product);
-                //     context.SaveChanges();
-                //     return RedirectToAction("Index");
-                // }
-
-
-                _dbContext.Add(product);
-                _dbContext.SaveChanges();
-
-                if (product.Image != null)
+                var existingProduct = await _dbContext.Products.FirstOrDefaultAsync(p => p.Id == product.Id);
+                if (existingProduct != null)
                 {
-                    string fileName = product.Id + Path.GetExtension(product.Image.FileName);
-                    string filePath = @"wwwroot\lib\Product\SmartPhone\" + fileName;
-
-                    var directoryLocation = Path.Combine(Directory.GetCurrentDirectory(), filePath);
-                    FileInfo fileInfo = new FileInfo(directoryLocation);
-                    if (fileInfo.Exists)
-                    {
-                        fileInfo.Delete();
-                    }
-
-                    var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
-                    using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
-                    {
-                        product.Image.CopyTo(fileStream);
-                    }
-                    var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
-                    product.ImageUrl = baseUrl + "/SmartPhone/" + fileName;
-                    product.ImageLocalPath = filePath;
+                    // Nếu sản phẩm đã tồn tại, trả về thông báo hoặc xử lý lỗi
+                    _response.Message = "Product already exists.";
+                    _response.StatusCode = HttpStatusCode.BadRequest;
+                    _response.IsSuccess = false;
                 }
                 else
                 {
-                    product.ImageUrl = "https://placehold.co/600x400";
-                }
-                _dbContext.Update(product);
-                _dbContext.SaveChanges();
+                    // // Xử lý thêm sản phẩm, cho phép các trường Size, Color, Brand và Category nullable
+                    // if (ModelState.IsValid)
+                    // {
+                    //     product.Id = Guid.NewGuid().ToString();
 
-                _response.StatusCode = HttpStatusCode.OK;
-                _response.IsSuccess = true;
-                _response.Result = product;
-                _response.Message = "Create product successfully";
+                    //     // Nếu người dùng không chọn, các thuộc tính này sẽ là null
+                    //     product.Size = product.Size?.Id != null ? context.Sizes.Find(product.Size.Id) : null;
+                    //     product.Color = product.Color?.Id != null ? context.Colors.Find(product.Color.Id) : null;
+                    //     product.Brand = product.Brand?.Id != null ? context.Brands.Find(product.Brand.Id) : null;
+                    //     product.Category = product.Category?.Id != null ? context.Categories.Find(product.Category.Id) : null;
+
+                    //     context.Products.Add(product);
+                    //     context.SaveChanges();
+                    //     return RedirectToAction("Index");
+                    // }
+
+                    _dbContext.Add(product);
+                    _dbContext.SaveChanges();
+
+                    if (product.Image != null)
+                    {
+                        string fileName = product.Id + Path.GetExtension(product.Image.FileName);
+                        string filePath = @"wwwroot\lib\Product\SmartPhone\" + fileName;
+
+                        var directoryLocation = Path.Combine(Directory.GetCurrentDirectory(), filePath);
+                        FileInfo fileInfo = new FileInfo(directoryLocation);
+                        if (fileInfo.Exists)
+                        {
+                            fileInfo.Delete();
+                        }
+
+                        var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
+                        using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
+                        {
+                            product.Image.CopyTo(fileStream);
+                        }
+                        var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
+                        product.ImageUrl = baseUrl + "/SmartPhone/" + fileName;
+                        product.ImageLocalPath = filePath;
+                    }
+                    else
+                    {
+                        product.ImageUrl = "https://placehold.co/600x400";
+                    }
+                    _dbContext.Update(product);
+                    _dbContext.SaveChanges();
+
+                    _response.Result = product;
+                    _response.Message = "Create product successfully";
+                }
             }
             catch (Exception ex)
             {
@@ -215,35 +226,43 @@ namespace App.Services.ProductAPI.Controllers.v1
         {
             try
             {
-                if (product.Image != null)
+                Product p = _dbContext.Products.FirstOrDefault(u => u.Id == product.Id);
+                if (p is null)
                 {
-                    if (!string.IsNullOrEmpty(product.ImageLocalPath))
-                    {
-                        var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), product.ImageLocalPath);
-                        FileInfo file = new FileInfo(oldFilePathDirectory);
-                        if (file.Exists)
-                        {
-                            file.Delete();
-                        }
-                    }
-                    string fileName = product.Id + Path.GetExtension(product.Image.FileName);
-                    string filePath = @"wwwroot\ProductImages\" + fileName;
-                    var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
-                    using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
-                    {
-                        product.Image.CopyTo(fileStream);
-                    }
-                    var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
-                    product.ImageUrl = baseUrl + "/ProductImages/" + fileName;
-                    product.ImageLocalPath = filePath;
+                    _response.StatusCode = HttpStatusCode.NotFound;
+                    _response.IsSuccess = false;
                 }
+                else
+                {
 
-                _dbContext.Update(product);
-                _dbContext.SaveChanges();
+                    if (product.Image != null)
+                    {
+                        if (!string.IsNullOrEmpty(product.ImageLocalPath))
+                        {
+                            var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), product.ImageLocalPath);
+                            FileInfo file = new FileInfo(oldFilePathDirectory);
+                            if (file.Exists)
+                            {
+                                file.Delete();
+                            }
+                        }
+                        string fileName = product.Id + Path.GetExtension(product.Image.FileName);
+                        string filePath = @"wwwroot\ProductImages\" + fileName;
+                        var filePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), filePath);
+                        using (var fileStream = new FileStream(filePathDirectory, FileMode.Create))
+                        {
+                            product.Image.CopyTo(fileStream);
+                        }
+                        var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.Value}{HttpContext.Request.PathBase.Value}";
+                        product.ImageUrl = baseUrl + "/ProductImages/" + fileName;
+                        product.ImageLocalPath = filePath;
+                    }
 
-                _response.StatusCode = HttpStatusCode.OK;
-                _response.IsSuccess = true;
-                _response.Message = "Product update successfully";
+                    _dbContext.Update(product);
+                    _dbContext.SaveChanges();
+
+                    _response.Message = "Product update successfully";
+                }
             }
             catch (Exception ex)
             {
@@ -260,22 +279,28 @@ namespace App.Services.ProductAPI.Controllers.v1
         {
             try
             {
-                Product product = _dbContext.Products.First(u => u.Id == id);
-                if (!string.IsNullOrEmpty(product.ImageLocalPath))
+                Product product = _dbContext.Products.FirstOrDefault(u => u.Id == id);
+                if (product == null)
                 {
-                    var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), product.ImageLocalPath);
-                    FileInfo file = new FileInfo(oldFilePathDirectory);
-                    if (file.Exists)
-                    {
-                        file.Delete();
-                    }
+                    _response.StatusCode = HttpStatusCode.NotFound;
+                    _response.IsSuccess = false;
                 }
-                _dbContext.Remove(product);
-                _dbContext.SaveChanges();
+                else
+                {
+                    if (!string.IsNullOrEmpty(product.ImageLocalPath))
+                    {
+                        var oldFilePathDirectory = Path.Combine(Directory.GetCurrentDirectory(), product.ImageLocalPath);
+                        FileInfo file = new FileInfo(oldFilePathDirectory);
+                        if (file.Exists)
+                        {
+                            file.Delete();
+                        }
+                    }
+                    _dbContext.Remove(product);
+                    _dbContext.SaveChanges();
 
-                _response.Message = "Product deleted successfully";
-                _response.StatusCode = HttpStatusCode.OK;
-                _response.IsSuccess = true;
+                    _response.Message = "Product deleted successfully";
+                }
             }
             catch (Exception ex)
             {
